# lib/api-zod/src/generated/types/account.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Account {
  id: number;
  name: string;
  /** checking, savings, credit, investment */
  type: string;
  balance: number;
  currency: string;
  /** @nullable */
  institution: string | null;
  /** @nullable */
  color: string | null;
  createdAt: Date;
  updatedAt: Date;
}