# lib/api-zod/src/generated/types/budget.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Budget {
  id: number;
  category: string;
  limit: number;
  spent: number;
  /** monthly or weekly */
  period: string;
  /** @nullable */
  color: string | null;
  createdAt: Date;
  updatedAt: Date;
}