# lib/api-zod/src/generated/types/createAccountBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateAccountBody {
  name: string;
  type: string;
  balance: number;
  currency: string;
  /** @nullable */
  institution?: string | null;
  /** @nullable */
  color?: string | null;
}