# lib/api-zod/src/generated/types/createBudgetBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateBudgetBody {
  category: string;
  limit: number;
  spent: number;
  period: string;
  /** @nullable */
  color?: string | null;
}