# lib/api-zod/src/generated/types/createGoalBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateGoalBody {
  name: string;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  targetDate?: Date | null;
  /** @nullable */
  category?: string | null;
  /** @nullable */
  color?: string | null;
}