# lib/api-zod/src/generated/types/createTransactionBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateTransactionBody {
  accountId: number;
  amount: number;
  type: string;
  category: string;
  description: string;
  date: Date;
  /** @nullable */
  notes?: string | null;
}