# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  totalBalance: number;
  totalIncome: number;
  totalExpenses: number;
  netSavings: number;
  accountCount: number;
  transactionCount: number;
  budgetCount: number;
  goalCount: number;
}