# lib/api-zod/src/generated/types/goal.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeloxFi API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Goal {
  id: number;
  name: string;
  targetAmount: number;
  currentAmount: number;
  /** @nullable */
  targetDate: Date | null;
  /** @nullable */
  category: string | null;
  /** @nullable */
  color: string | null;
  createdAt: Date;
  updatedAt: Date;
}